+++ src/convtools/contrib/models/casters/__init__.py
@@ -1,5 +1,6 @@
 """Exposes casters to be made publicly available"""
 from .casters import (
+    Bool,
     Custom,
     CustomUnsafe,
     DateFromStr,